requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
dj-stripe==2.9.1
Django==5.2.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
packaging==25.0
pillow==11.1.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-environ==0.4.54
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
stripe==11.6.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0